add group chat

+++ db/DeleteClient.go
@@ -14,13 +14,25 @@ func DeleteClient(user *models.User, hashname string) error {
 	if user.Hashname == hashname {
 		return errors.New("Can't delete user from function delete client")
 	}
+	clientid := GetClientId(id, hashname)
+	if clientid < 0 {
+		return errors.New("Client id undefined")
+	}
 	_, err := settings.DB.Exec(
 		"DELETE FROM Chat WHERE IdUser=$1 AND IdClient=$2", 
 		id,
-		GetClientId(id, hashname),
+		clientid,
 	)
 	if err != nil {
 		panic("exec 'deleteclient.chat' failed")
+	}
+	_, err = settings.DB.Exec(
+		"DELETE FROM GlobalChat WHERE IdUser=$1 AND Founder=$2", 
+		id,
+		hashname,
+	)
+	if err != nil {
+		panic("exec 'deleteclient.globalchat' failed")
 	}
 	_, err = settings.DB.Exec(
 		"DELETE FROM Client WHERE IdUser=$1 AND Hashname=$2",
